+++ CHANGES.md
@@ -22,6 +22,7 @@ API
   * `Call_provers.print_prover_result` now takes an additional argument
     `~json_model` to indicate whether counterexamples are printed using JSON :x:
   * Counterexamples: indices of array are now model_value. :x:
+  * ITP constructor Task now contain the location of the goal :x:
 
 Transformations
   * `apply`/`rewrite` behaves better in presence of `let`;